pyproject.toml
[project]
name = "pyqttier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jacob Brunson", email = "github@jacobbrunson.com" }
]
requires-python = ">=3.7"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pydantic>=2.5.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=23.3.0",
]
